+++ src/Collection.tsx
@@ -1,0 +1,12 @@
+import React from "react";
+import {View, Text} from "react-native";
+
+export default class Collection extends React.Component {
+  render() {
+    return(
+      <View>
+        <Text>Collection</Text>
+      </View>
+    );
+  }
+}
