+++ src/Main.tsx
@@ -1,5 +1,5 @@
 import React, {FC} from "react";
-import {Button, Text, View} from "react-native";
+import {Button, Text, View, StyleSheet} from "react-native";
 import {Events, Props, State} from "./main/interface";
 import {Dispatch} from "redux";
 import {notTakeCare, takeCare} from "./main/actions";
@@ -7,7 +7,7 @@ import {connect} from "react-redux";
 import { Pet } from "./components/Pet";
 
 export const Main: FC<Props> = props => (
-  <View>
+  <View style={styles.container}>
     <Text>{props.stage}</Text>
     <Text>{props.health}</Text>
     <Pet stage={props.stage}/>
@@ -15,6 +15,15 @@ export const Main: FC<Props> = props => (
     <Button title={'世話をしない'} onPress={props.notTakeCare}/>
   </View>
 );
+
+const styles = StyleSheet.create({
+  container: {
+    flex: 1,
+    justifyContent: 'center',
+    alignItems: 'center',
+    backgroundColor: '#F5FCFF',
+  },
+});
 
 export const mapStateToProps = (state: State) => state;
 export const mapDispatchToProps = (dispatch: Dispatch): Events => ({
